Moved remaining reasoning methods from CausaloidGraph into the default implementation in causable_graph_explaining protocol
